backend/requirements.txt
fastapi
uvicorn
redis 
groq
python-multipart 
yt-dlp 
pymupdf
python-dotenv
YouTubeTranscriptApi